requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
python-dotenv==1.1.1
